move blueprint example() out of [cfg(test)]
