+++ nexus/reconfigurator/planning/src/lib.rs
@@ -116,6 +116,7 @@
 //! updates, etc.
 
 pub mod blueprint_builder;
+pub mod example;
 mod ip_allocator;
 pub mod planner;
 pub mod system;
